requirements.txt
zeroconf==0.135.*
requests==2.*
textual==0.81.*
Pillow==11.*
